Don't install self :package:
When running `renv::install` exclude any package that matches the name of the analysis directory. This ensures that if the project is setup as an R package with tests, `renv` will not infer itself as a package to be installed which will fail (unless it happens to be on CRAN).