Handle package exclusion more robustly
In the case that the analysis directory is an R package with tests, it will introduce the issue where `testthat.R` calls `library(my_pkg)` where `my_pkg` is the local package and likely cannot be installed by renv. I am avoiding this by excluding the package name found in DESCRIPTION files anywhere in the analysis directory.